+++ src/main/java/com/maxmind/db/Decoder.java
@@ -439,6 +439,17 @@ final class Decoder {
                 InvocationTargetException e) {
             throw new DeserializationException("Error creating object: " + e.getMessage(), e);
         }
+        catch (IllegalArgumentException e){
+            StringBuilder sbErrors = new StringBuilder();  
+            for (String key : parameterIndexes.keySet()) {
+                int index = parameterIndexes.get(key);
+                if (!parameters[index].getClass().isAssignableFrom( parameterTypes[index])) {
+                    sbErrors.append(" argument type mismatch in " + key + " MMDB Type: " + parameters[index].getClass().getCanonicalName()
+                    + " Java Type: " +parameterTypes[index].getCanonicalName());
+                }              
+            }
+            throw new DeserializationException("Error creating object of type: " + cls.getSimpleName() + " - " + sbErrors.toString(), e);
+        }
     }
 
     private static <T> Constructor<T> findConstructor(Class<T> cls)