+++ src/test/java/com/maxmind/db/ReaderTest.java
@@ -568,6 +568,61 @@ public class ReaderTest {
     }
 
     @Test
+    public void testDecodeWrongTypeWithConstructorException() throws IOException {
+        this.testReader = new Reader(getFile("GeoIP2-City-Test.mmdb"));
+        DeserializationException ex = assertThrows(DeserializationException.class,
+            () -> this.testReader.get( InetAddress.getByName("2.125.160.216"),
+            TestModelSubdivisionsWithUnknownException.class));
+        
+        assertThat(ex.getMessage(), containsString("Error getting record for IP /2.125.160.216 -  Error creating object"));
+    }
+
+    static class TestModelSubdivisionsWithUnknownException {
+        List<TestModelSubdivision> subdivisions;
+
+        @MaxMindDbConstructor
+        public TestModelSubdivisionsWithUnknownException (
+            @MaxMindDbParameter(name="subdivisions")
+            List<TestModelSubdivision> subdivisions
+        ) throws Exception{
+            throw new Exception();
+        }
+    }
+
+    @Test
+    public void testDecodeWrongTypeWithWrongArguments() throws IOException {
+        this.testReader = new Reader(getFile("GeoIP2-City-Test.mmdb"));
+        DeserializationException ex = assertThrows(DeserializationException.class,
+            () -> this.testReader.get( InetAddress.getByName("2.125.160.216"),
+            TestWrongModelSubdivisions.class));
+        assertThat(ex.getMessage(), containsString("Error getting record for IP"));
+    }
+
+    static class TestWrongModelSubdivisions {
+        List<TestWrongModelSubdivision> subdivisions;
+
+        @MaxMindDbConstructor
+        public TestWrongModelSubdivisions (
+            @MaxMindDbParameter(name="subdivisions")
+            List<TestWrongModelSubdivision> subdivisions
+        ) {
+            this.subdivisions = subdivisions;
+        }
+    }
+
+    static class TestWrongModelSubdivision {
+        Integer uint16Field;
+        @MaxMindDbConstructor
+        public TestWrongModelSubdivision(
+            @MaxMindDbParameter(name="iso_code")
+            Integer uint16Field
+            ) {
+            this.uint16Field = uint16Field;
+            ;
+        }
+    }
+
+    @Test
     public void testDecodeConcurrentHashMap() throws IOException {
         this.testReader = new Reader(getFile("GeoIP2-City-Test.mmdb"));
 
@@ -775,7 +830,7 @@ public class ReaderTest {
                 TestModelInvalidMap.class
             )
         );
-        assertEquals("Map keys must be strings.", ex.getMessage());
+        assertEquals("Error getting record for IP /2.125.160.216 -  Map keys must be strings.", ex.getMessage());
     }
 
     static class TestModelInvalidMap {