+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 CHANGELOG
 =========
-
 2.1.0
 ------------------
 
-* Messages for `DeserializationException` have been improved and the
-  cause is included, if any.
+* Messages for `DeserializationException` have been improved and the cause
+ is included, if any. Moreover, the message provides detail about the involved
+ types, if the exeption is caused by an `IllegalArgumentException`.
 
 2.0.0 (2020-10-13)
 ------------------